+++ src/main/java/org/truffleruby/language/literal/StringLiteralNode.java
@@ -20,11 +20,14 @@ package org.truffleruby.language.literal;
 
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.object.DynamicObject;
+import org.truffleruby.Layouts;
 import org.truffleruby.core.rope.Rope;
-import org.truffleruby.core.string.StringOperations;
 import org.truffleruby.language.RubyNode;
+import org.truffleruby.language.objects.AllocateObjectNode;
 
 public class StringLiteralNode extends RubyNode {
+
+    @Child AllocateObjectNode allocateNode = AllocateObjectNode.create();
 
     private final Rope rope;
 
@@ -35,7 +38,7 @@ public class StringLiteralNode extends RubyNode {
 
     @Override
     public DynamicObject execute(VirtualFrame frame) {
-        return StringOperations.createString(getContext(), rope);
+        return allocateNode.allocate(coreLibrary().getStringClass(), Layouts.STRING.build(false, false, rope));
     }
 
 }