+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Bug fixes:
 * Fixed `Process.kill(signal, Process.pid)` when the signal is trapped as `:IGNORE` (#1702).
 * Fixed `Addrinfo.new(String)` to reliably find the address family (#1702).
 * Fixed argument checks in `BasicSocket#setsockopt` (#1460).
+* Fixed `ObjectSpace.trace_object_allocations` (#1456).
 
 Compatibility: